requirements.txt
absl-py==0.7.1
astor==0.8.0
boto3==1.9.175
botocore==1.12.175
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docutils==0.14
Flask==1.0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.8.0
python-dateutil==2.8.0
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wrapt==1.11.2
